suggest 식단 추천 기능 완료

+++ src/main/java/com/witheat/WithEatServer/Domain/Dto/request/MemberHeightRequestDto.java
@@ -10,6 +10,6 @@ import java.time.LocalDate;
 @NoArgsConstructor
 @AllArgsConstructor
 public class MemberHeightRequestDto {
-    private Long height;
+    private int height;
     private LocalDate height_date;
 }
